test: move BasePath and State routing ITs to test-default
Both are plain Java views (no TS): BasePath checks a route with a
'view/' prefix resolves, State checks React is enabled and present in
package.json. Moved from test-react-router to test-default (React home),
JUnit 6 + @TestFor, deleted from source in the same commit.
