+++ flow-tests/test-default/src/main/java/com/vaadin/flow/test/routing/BasePathView.java
@@ -13,28 +13,16 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.flow;
+package com.vaadin.flow.test.routing;
 
-import org.junit.Assert;
-import org.junit.Test;
+import com.vaadin.flow.component.html.Div;
+import com.vaadin.flow.component.html.Span;
+import com.vaadin.flow.router.Route;
 
-import com.vaadin.flow.component.html.testbench.SpanElement;
-import com.vaadin.flow.testutil.ChromeBrowserTest;
+@Route("view/com.vaadin.flow.BasePathView")
+public class BasePathView extends Div {
 
-public class BasePathIT extends ChromeBrowserTest {
-    @Override
-    protected String getTestPath() {
-        return "/view/view/com.vaadin.flow.BasePathView";
+    public BasePathView() {
+        add(new Span("BasePathView"));
     }
-
-    @Test
-    public void navigationTo_routeWithBasePath_succeeds() {
-        open();
-
-        waitForDevServer();
-
-        Assert.assertTrue(
-                $(SpanElement.class).withText("BasePathView").exists());
-    }
-
 }
