+++ flow-tests/test-default/src/main/java/com/vaadin/flow/test/routing/StateView.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.flow;
+package com.vaadin.flow.test.routing;
 
 import java.io.File;
 import java.io.IOException;
